No longer quit if reading STDIN returns -1
There is some unexpected (at least by me and a couple others who
looked into it) behavior going on where if a user's system is
configured a certain way (seems like most likely when using a specific
version of bash) then if lein-test-refresh runs some code that causes
bash to be invoked, then STDIN starts returning -1 when reading from
it.

Prior to this commit, that would caused lein-test-refresh to call
(System/exit 0) and unexpectedly quit.

Now, lein-test-refesh will just print a message and hope the users
sees it. This is probably going to be confusing. This message will
appear when someone hits CTRL-D as well.